refactored the test suite code
